backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
pydantic==2.11.3
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
google-cloud-texttospeech==2.25.0
